# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateProjectRequest:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      max_team_size:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - title
    type: object
...
  dto.ProjectApplicationResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
      user_id:
        type: string
    type: object
  dto.ProjectDetailResponse:
    properties:
      applications:
        items:
          $ref: '#/definitions/dto.ProjectApplicationResponse'
        type: array
      can_manage:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_member:
        type: boolean
      is_owner:
        type: boolean
      is_public:
        type: boolean
      max_team_size:
        type: integer
      member_count:
        type: integer
      members:
        items:
          $ref: '#/definitions/dto.ProjectMemberResponse'
        type: array
      owner:
        $ref: '#/definitions/dto.UserResponse'
      owner_id:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.ProjectMemberResponse:
    properties:
      joined_at:
        type: string
      role:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
      user_id:
        type: string
    type: object
  dto.ProjectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      max_team_size:
        type: integer
      member_count:
        type: integer
      owner:
        $ref: '#/definitions/dto.UserResponse'
      owner_id:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@synergyconnect.com
    name: API Support
  description: Университетская социальная сеть с полным набором функций
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://synergyconnect.com/terms
  title: SynergyConnect API
  version: 1.0.0
paths:
...
  /projects:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для создания проекта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создание проекта
      tags:
      - projects
  /projects/{id}:
    get:
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectDetailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получение проекта
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"